pia adicionei o projeto ae, mais não mexa no carrinho, nem em produto, ainda tem uns bugs doidos la, e nao compra 100%

+++ src/main/java/model/entities/Carrinho.java
@@ -8,7 +8,7 @@ public class Carrinho implements EntityBase {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-    private long valorTotal;
+    private double valorTotal;
     private long quantidade;
     
     @ManyToOne
@@ -38,11 +38,11 @@ public class Carrinho implements EntityBase {
         this.quantidade = quantidade;
     }
 
-    public long getValorTotal() {
+    public double getValorTotal() {
         return valorTotal;
     }
 
-    public void setValorTotal(long valorTotal) {
+    public void setValorTotal(double valorTotal) {
         this.valorTotal = valorTotal;
     }
